pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "netifaces>=0.11.0",
    "psutil>=7.2.1",
    "pyyaml>=6.0",
]
